Writing exact package version into web/version.txt.
The update-check is not run automatically and the
update-button is not shown on Debian.

If the update-check is triggered manually, it expects
version.txt to contain something, therefore on Debian
it contained '0'.

As it does not matter what value is in there,
there is no gain in writing an inaccurate one either.

Therefore, let's write the exact package version in it.

Note: version.txt is by default exposed to the web via
http://localhost:19999/version.txt. Anyone who can access
netdata can also see what exact version you're running.

Signed-off-by: Daniel Baumann <daniel.baumann@progress-linux.org>